set up user manegment class to start creating login component

+++ BachiraApp/src/app/classes/UserManagement.ts
@@ -1,30 +1,48 @@
 import { User } from "./User";
 
 export class UserManagement {
-    static users_data: any = JSON.parse(localStorage.getItem('user_data') + " ");
-    static users: User[] = this.users_data;
+    //#region User array
+    private static users_data: any = JSON.parse(localStorage.getItem('user_data') + "");
+    private static users: User[] = this.users_data;
+    //#endregion
 
-    static Update(user: User) {
-        if (this.users == null)
-            this.users = [];
+    public static Update(user: User) {
+        if (this.users == null) this.users = [];
         
-        for (let i = 0; i < this.users.length; i++) {
-            const user_ = this.users[i];
-            if (user.email == user_.email) {
-                return false;
-            }
-        }
+        if (this.Exists(user)) return false;
 
         this.users.push(user);
         this.users_data = JSON.stringify(this.users);
         localStorage.setItem('user_data', this.users_data);
 
-
         return true;
-        
     }
 
-    static TEST() {
-        console.log(localStorage.getItem('user_data'));
+    public static IsValidUser(user: User) {
+        if (!this.Exists(user)) return false;
+        let userOfEmail: User | undefined = this.users.find((user_)=> { return user_.email == user.email })
+
+        return userOfEmail?.password == user.password && userOfEmail != undefined;
+    }
+
+    public static PrintUsers() {
+        console.log(this.users_data);
+    }
+
+    public static GetFullUser(email: string, password:string) {
+        return this.users.find((user_)=> {
+            return user_.email == email && user_.password == password;
+        })
+    }
+
+    private static Exists(user: User) {
+        for (let i = 0; i < this.users.length; i++) {
+            const user_ = this.users[i];
+            if (user.email == user_.email) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
